+++ wrapper/IO/BrokenParser.pypp.cpp
@@ -79,6 +79,8 @@ namespace bp = boost::python;
 
 SireIO::BrokenParser __copy__(const SireIO::BrokenParser &other){ return SireIO::BrokenParser(other); }
 
+#include "Helpers/copy.hpp"
+
 #include "Qt/qdatastream.hpp"
 
 #include "Helpers/str.hpp"
@@ -260,9 +262,9 @@ void register_BrokenParser_class(){
         
         }
         BrokenParser_exposer.staticmethod( "typeName" );
-        BrokenParser_exposer.def( "__copy__", &__copy__);
-        BrokenParser_exposer.def( "__deepcopy__", &__copy__);
-        BrokenParser_exposer.def( "clone", &__copy__);
+        BrokenParser_exposer.def( "__copy__", &__copy__<SireIO::BrokenParser>);
+        BrokenParser_exposer.def( "__deepcopy__", &__copy__<SireIO::BrokenParser>);
+        BrokenParser_exposer.def( "clone", &__copy__<SireIO::BrokenParser>);
         BrokenParser_exposer.def( "__rlshift__", &__rlshift__QDataStream< ::SireIO::BrokenParser >,
                             bp::return_internal_reference<1, bp::with_custodian_and_ward<1,2> >() );
         BrokenParser_exposer.def( "__rrshift__", &__rrshift__QDataStream< ::SireIO::BrokenParser >,