+++ wrapper/IO/PerturbationsTemplate.pypp.cpp
@@ -75,6 +75,8 @@ namespace bp = boost::python;
 
 SireIO::PerturbationsTemplate __copy__(const SireIO::PerturbationsTemplate &other){ return SireIO::PerturbationsTemplate(other); }
 
+#include "Helpers/copy.hpp"
+
 #include "Qt/qdatastream.hpp"
 
 const char* pvt_get_name(const SireIO::PerturbationsTemplate&){ return "SireIO::PerturbationsTemplate";}
@@ -645,9 +647,9 @@ void register_PerturbationsTemplate_class(){
         
         }
         PerturbationsTemplate_exposer.staticmethod( "typeName" );
-        PerturbationsTemplate_exposer.def( "__copy__", &__copy__);
-        PerturbationsTemplate_exposer.def( "__deepcopy__", &__copy__);
-        PerturbationsTemplate_exposer.def( "clone", &__copy__);
+        PerturbationsTemplate_exposer.def( "__copy__", &__copy__<SireIO::PerturbationsTemplate>);
+        PerturbationsTemplate_exposer.def( "__deepcopy__", &__copy__<SireIO::PerturbationsTemplate>);
+        PerturbationsTemplate_exposer.def( "clone", &__copy__<SireIO::PerturbationsTemplate>);
         PerturbationsTemplate_exposer.def( "__rlshift__", &__rlshift__QDataStream< ::SireIO::PerturbationsTemplate >,
                             bp::return_internal_reference<1, bp::with_custodian_and_ward<1,2> >() );
         PerturbationsTemplate_exposer.def( "__rrshift__", &__rrshift__QDataStream< ::SireIO::PerturbationsTemplate >,