+++ doc/source/changelog.rst
@@ -21,6 +21,8 @@ organisation on `GitHub <https://github.com/openbiosim/sire>`__.
 
 * Set appropriate default OpenMM box vectors when a system has no periodic space.
 
+* Added :func:`Sire::IO::setCoordinates` function to set atom coordinates of an entire system.
+
 `2025.2.0 <https://github.com/openbiosim/sire/compare/2025.1.0...2025.2.0>`__ - October 2025
 --------------------------------------------------------------------------------------------